Adding test coverage for the new way to defining generic classes
